requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
curl-cffi==0.5.9
Django==4.2.3
django-ajax==0.3.0
django-bootstrap-modal-forms==3.0.4
django-cors-headers==4.2.0
django-cors-middleware==1.5.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
duckduckgo-search==3.9.4
exceptiongroup==1.2.2
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
g4f==0.3.2.6
greenlet==3.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
Pillow==10.1.0
platformdirs==3.11.0
playwright==1.39.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pyee==11.0.1
PyExecJS==1.5.1
pyjsparser==2.7.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
sqlparse==0.4.4
tiktoken==0.5.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
waitress==2.1.2
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
yarl==1.9.2
